Move magictypes to types file and fixed typos in spells
